Enemy free hit on player flee
Moved finalizing the player move to a new 'MoveTo' function.
Renamed cardinal move functions to 'AttempMoveX', which is also a more apt name since these functions may be called without being a possible move.
Provided enemy a free attack if _currentBattle is not null, and Disposed of battle when fleeing.